Consolidate Stimulus in search-status_controller.js
